Create FirebaseTickerFragment to share code
